bashrc - Add bashrc2 to contain common config
Unfortunately there's going to be a manual installation step to source this
bashrc2 file in the main bashrc. However, there is often a lot of default
stuff in the bashrc that's added automatically that I don't want to either
override or version control. It would be good to add something to the
installation script (in a re-usable manner) to allow automatic appending of
the source command to the existing .bashrc (obviously checking if it exists or
not first, as running ./install multiple times shouldn't append multiple times).